Fix float conversion warnings in LinuxPowerProbe.cpp
- Change chargeNowWh, chargeFullWh, chargeDesignWh from uint64_t to double
  in PowerTypes.h since battery capacity values are fractional (e.g., 45.5 Wh)
- Use floating-point division (/ 1000000.0) instead of integer division
  when converting µWh to Wh to preserve precision
- Fixes -Wfloat-conversion and -Wimplicit-int-float-conversion warnings
  that were causing Linux CI build failures

+++ src/Platform/Linux/LinuxPowerProbe.cpp
@@ -158,9 +158,9 @@ void LinuxPowerProbe::readBattery(PowerCounters& counters, const std::string& ba
         // Try energy first (Wh)
         if (std::filesystem::exists(batteryPath + "/energy_now"))
         {
-            counters.chargeNowWh = readSysfsUInt64(batteryPath + "/energy_now") / 1000000; // µWh to Wh
-            counters.chargeFullWh = readSysfsUInt64(batteryPath + "/energy_full") / 1000000;
-            counters.chargeDesignWh = readSysfsUInt64(batteryPath + "/energy_full_design") / 1000000;
+            counters.chargeNowWh = static_cast<double>(readSysfsUInt64(batteryPath + "/energy_now")) / 1000000.0; // µWh to Wh
+            counters.chargeFullWh = static_cast<double>(readSysfsUInt64(batteryPath + "/energy_full")) / 1000000.0;
+            counters.chargeDesignWh = static_cast<double>(readSysfsUInt64(batteryPath + "/energy_full_design")) / 1000000.0;
         }
         // Fall back to charge (Ah) - need voltage to convert to Wh
         else if (std::filesystem::exists(batteryPath + "/charge_now"))
